fix: Update URLText Searcher Processor
The URLTextSearcher processor needed the "url" parameter which I had
removed due to an error message. I added it back, but with a different
URL, and changed the Regex in the "re_pattern" parameter.
I tested the recipe locally, and it was able to grab the DMG for further
processing and packaging.