refactor: Overhaul The Docker Recipes
I have overhauled the download and packaging recipes with assistance from
the smithjw-recipes Autopkg Recipe repo. This appears to be the most up
to date in terms of methodology for getting Docker to install properly
on machines.
I just needed to adjust the architecture to work with Apple Silicon
devices, as we are trying to phase out the Intel devices.